tried some different footer styles
Decided that the footer should probably be just like the original websites with the inclusion of the social media buttons being placed there now. They will also be placed in the contact page. While it would be a good idea to have all of the contact information on the first page, if that is done it shouldn't be that footer for every page