worked on mural text for the about page
Added full mural image and a prototype of the history section. discovered that the problem is that instead of redesigning the mural details png to get a better font size,
the issue is actually that a png's quality isn't very good outside its original size. Tried looking for websites
that are able to convert a png into html and css. But the main recommendation of fronty was under maintanence and the rest sucked. Tried altering sucky code to make it work. Made a strategy to deal with this problem.