nearly completed menu page
compressed all images so that the load time will be normal, the menu page ended up being as large as the about page after the compressions. Some quality was lost but not much. Organized all images in website to subfolders and removed unused images. Added all but two of the menu images, still missing two of the holiday breads.